updated two tests:
"update element with dynamic props" now correctly omits the `value` attribute from the expected markup - because we're now updating the property, which is not the same as updating the attribute.

"elements with falsy values" now uses an attribute for the tests, rather than `value`, which is a property.

perhaps still to do: add a test for handling of falsy property values.