Fixed bad result of loop variable.
In patchElement, there is a var i initialized at the top of the function for use in while loops. Then at the end of the function a for loop had another var i use. Variables declared in loops are automatically hoisted to the top of a function. So, in this case the browser will discard the var and reuse the previously declared i. Since the first declaration of i is for numberical calculations, this causes a type error for the for loop, since the value here is of type string. Renaming the variable in the for loop fixes this problems and avoids any unexpected results caused by a possible number value for a previous loop getting passed to the for loop.